# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个游戏管理系统的API文档，包含前端用户、后台管理、游戏管理等功能",
        "title": "游戏管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8801",
    "basePath": "/",
    "paths": {
...
        "/v1/admin/games": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取游戏列表，支持多种条件筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-游戏管理"
                ],
                "summary": "获取游戏列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "游戏中文名称（模糊查询）",
                        "name": "name_zh",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "游戏英文名称（模糊查询）",
                        "name": "name_en",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "厂商ID（精确查询）",
                        "name": "publisher_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "工作室ID（精确查询）",
                        "name": "studio_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游戏类型编码（精确查询）",
                        "name": "type_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游戏平台编码（精确查询）",
                        "name": "platform_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游戏语言编码（精确查询）",
                        "name": "language_code",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最低评分",
                        "name": "min_rating",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最高评分",
                        "name": "max_rating",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最低价格",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最高价格",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态（0停用 1正常）",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.PageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "records": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.GameRes"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的游戏记录，包括基本信息和关联数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-游戏管理"
                ],
                "summary": "创建游戏",
                "parameters": [
                    {
                        "description": "游戏创建请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GameReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/games/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据游戏ID获取游戏的详细信息，包括所有关联数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-游戏管理"
                ],
                "summary": "获取游戏详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "游戏ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GameRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新游戏信息，包括基本信息和关联数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-游戏管理"
                ],
                "summary": "更新游戏",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "游戏ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "游戏更新请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GameReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据游戏ID删除游戏及其所有关联数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-游戏管理"
                ],
                "summary": "删除游戏",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "游戏ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.PageResponse": {
            "type": "object",
            "properties": {
                "current": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "pages": {
                    "description": "总页数",
                    "type": "integer"
                },
                "records": {
                    "description": "数据列表"
                },
                "size": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据集"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "dto.GameCoverReq": {
            "type": "object",
            "required": [
                "cover_url"
            ],
            "properties": {
                "cover_id": {
                    "description": "封面ID，更新时需要",
                    "type": "integer"
                },
                "cover_url": {
                    "description": "封面图片URL，必填",
                    "type": "string"
                },
                "is_main": {
                    "description": "是否主封面，必填",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "sort": {
                    "description": "排序，可选",
                    "type": "integer"
                }
            }
        },
        "dto.GameCoverRes": {
            "type": "object",
            "properties": {
                "cover_id": {
                    "description": "封面ID",
                    "type": "integer"
                },
                "cover_url": {
                    "description": "封面图片URL",
                    "type": "string"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "is_main": {
                    "description": "是否主封面",
                    "type": "integer"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                }
            }
        },
        "dto.GameLanguageInfo": {
            "type": "object",
            "properties": {
                "dict_code": {
                    "description": "字典编码",
                    "type": "integer"
                },
                "dict_label": {
                    "description": "字典标签",
                    "type": "string"
                },
                "dict_value": {
                    "description": "字典值",
                    "type": "string"
                }
            }
        },
        "dto.GamePlatformInfo": {
            "type": "object",
            "properties": {
                "dict_code": {
                    "description": "字典编码",
                    "type": "integer"
                },
                "dict_label": {
                    "description": "字典标签",
                    "type": "string"
                },
                "dict_value": {
                    "description": "字典值",
                    "type": "string"
                }
            }
        },
        "dto.GameReq": {
            "type": "object",
            "required": [
                "name_zh",
                "status"
            ],
            "properties": {
                "covers": {
                    "description": "封面和截图",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GameCoverReq"
                    }
                },
                "demo_video": {
                    "description": "游戏演示视频链接，可选",
                    "type": "string"
                },
                "description": {
                    "description": "游戏介绍，可选",
                    "type": "string"
                },
                "game_id": {
                    "description": "游戏ID，更新时需要，创建时忽略",
                    "type": "integer"
                },
                "language_codes": {
                    "description": "游戏语言编码数组",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name_en": {
                    "description": "游戏英文名称，可选",
                    "type": "string"
                },
                "name_zh": {
                    "description": "游戏中文名称，必填",
                    "type": "string"
                },
                "platform_codes": {
                    "description": "游戏平台编码数组",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "description": "游戏价格，可选",
                    "type": "number"
                },
                "publisher_id": {
                    "description": "游戏厂商ID，可选",
                    "type": "integer"
                },
                "rating": {
                    "description": "游戏评分，可选",
                    "type": "number"
                },
                "release_date": {
                    "description": "游戏发布日期，可选",
                    "type": "string"
                },
                "screenshots": {
                    "description": "游戏截图数组",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GameScreenshotReq"
                    }
                },
                "shutdown_date": {
                    "description": "游戏停服日期，可选",
                    "type": "string"
                },
                "size": {
                    "description": "游戏大小，可选",
                    "type": "string"
                },
                "status": {
                    "description": "状态，必填，只能是0或1",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "studio_id": {
                    "description": "游戏工作室ID，可选",
                    "type": "integer"
                },
                "type_codes": {
                    "description": "关联数据",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dto.GameRes": {
            "type": "object",
            "properties": {
                "comment_count": {
                    "description": "游戏评论数量",
                    "type": "integer"
                },
                "covers": {
                    "description": "游戏封面",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GameCoverRes"
                    }
                },
                "create_by": {
                    "description": "创建者",
                    "type": "string"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "demo_video": {
                    "description": "游戏演示视频链接",
                    "type": "string"
                },
                "description": {
                    "description": "游戏介绍",
                    "type": "string"
                },
                "game_id": {
                    "description": "游戏ID",
                    "type": "integer"
                },
                "languages": {
                    "description": "游戏语言信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GameLanguageInfo"
                    }
                },
                "like_count": {
                    "description": "游戏点赞数量",
                    "type": "integer"
                },
                "name_en": {
                    "description": "游戏英文名称",
                    "type": "string"
                },
                "name_zh": {
                    "description": "游戏中文名称",
                    "type": "string"
                },
                "platforms": {
                    "description": "游戏平台信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GamePlatformInfo"
                    }
                },
                "price": {
                    "description": "游戏价格",
                    "type": "number"
                },
                "publisher_id": {
                    "description": "游戏厂商ID",
                    "type": "integer"
                },
                "publisher_name": {
                    "description": "关联数据",
                    "type": "string"
                },
                "rating": {
                    "description": "游戏评分",
                    "type": "number"
                },
                "release_date": {
                    "description": "游戏发布日期",
                    "type": "string"
                },
                "screenshots": {
                    "description": "游戏截图",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GameScreenshotRes"
                    }
                },
                "shutdown_date": {
                    "description": "游戏停服日期",
                    "type": "string"
                },
                "size": {
                    "description": "游戏大小",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                },
                "studio_id": {
                    "description": "游戏工作室ID",
                    "type": "integer"
                },
                "studio_name": {
                    "description": "工作室名称",
                    "type": "string"
                },
                "types": {
                    "description": "游戏类型信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GameTypeInfo"
                    }
                },
                "update_by": {
                    "description": "更新者",
                    "type": "string"
                },
                "update_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "dto.GameScreenshotReq": {
            "type": "object",
            "required": [
                "screenshot_url"
            ],
            "properties": {
                "screenshot_id": {
                    "description": "截图ID，更新时需要",
                    "type": "integer"
                },
                "screenshot_url": {
                    "description": "截图URL，必填",
                    "type": "string"
                },
                "sort": {
                    "description": "排序，可选",
                    "type": "integer"
                }
            }
        },
        "dto.GameScreenshotRes": {
            "type": "object",
            "properties": {
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "screenshot_id": {
                    "description": "截图ID",
                    "type": "integer"
                },
                "screenshot_url": {
                    "description": "截图URL",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                }
            }
        },
        "dto.GameTypeInfo": {
            "type": "object",
            "properties": {
                "dict_code": {
                    "description": "字典编码",
                    "type": "integer"
                },
                "dict_label": {
                    "description": "字典标签",
                    "type": "string"
                },
                "dict_value": {
                    "description": "字典值",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}